Minimum build time too
Some distributions have patched previous OpenSSL and not provided any
indication in the version string that OpenSSL has been fixed.  Allow
administrators to choose instead to specify a minimum build time, which
should match the output from `openssl version -b`, which the runtime
must satisfy.

There are many assumptions inherent in "it was built at or after this
date, so therefore it must be good"; this new option is merely making
the best of a bad situation.

Note that I have an OpenSSL using "PDT", which `strptime()` handles, and
another using "EDT" which `strptime()` does not, so in the interests of
simplicity we ignore timezones.  Thus there is some error marging in
these values.  Really, the difference of a few hours, when using a
_build_ _timestamp_ shouldn't matter -- admins are after "on or after
date X", and _should_ be getting the exact same output they get from
`openssl version -b`, or something rather later.

So, hold our noses, this is good enough to achieve the goal, which is
working around missing version information for locally-applied patches.